write extra byte after command per sd spec
This effectively reduces power consumption from 15ma to 250ua between writes.
the line of code was commented out with a warn ("may cause errors") message
so this might need to be removed, but it worked on the Sandisk Ultra during
my testing. Power savings are significant so it should be worth it.